requirements.txt
aiogram==3.4.1
pika==1.3.2
python-dotenv==1.0.0